refactoring: moved 'lines_modified_by_commit' to 'Project' class
